fix URL parsing for cache key generation
As part of refactoring in #190 the URL parsing has been altered, s.t.
we pass a concatenation of host and path to wp_parse_url(). This results
in the actual host not being parsed at all which raises a PHP warning.

Prefix the URL with double slashes (the protocol is not of interest
here) to fix the parsing.