narrow down redis info scope and switch to "used_memory_dataset"
There is no need to fetch full server info, we only read from the memory
section. Also, "used_memory" might be misleading as this represents the
total memory allocation of the server instance. It may be confusing if
there are several megabyes allocated after flushing the cache, so
"used_memory_dataset" which subtracts the internal overhead is closer to
what a user might expect to see here.