attach flush button script to DOMContentLoaded (#312) (#326)
Depending on the actual rendering sequence, users might run into a race
condition where the admin bar element is not rendered at the time of
script execution.

Move the IIFE logic into a DOMContentLoaded event listener and add a
check for the existence of the admin bar element to prevent errors.